Move pointer generator to a separate file
